requirements.txt
customtkinter>=5.0.0
python-telegram-bot>=20.0
boto3>=1.26.0
paramiko>=3.0.0
google-api-python-client>=2.90.0
cryptography>=41.0.0
